chore : filterList 를 constants 폴더로 이동(#7)
