+++ src/components/FilterButtons.jsx
@@ -1,13 +1,6 @@
 import React from "react";
 import { Button } from "./Button";
-
-const FILTER_LABELS = {
-  category: "카테고리",
-  priority: "중요도",
-  share: "공유",
-  repeat: "반복",
-  latest: "최신순",
-};
+import { FILTER_LABELS } from "../constants/FILTER_LIST";
 
 const FilterButtons = ({ onFilterToggle }) => {
   return (
